requirements.txt
# RAG dependencies
chromadb>=0.4.22
langchain>=0.1.11
langchain-community>=0.0.27
langchain-ollama>=0.1.1
langchain-core>=0.1.30
python-dotenv>=1.0.1
pypdf>=4.0.2
pydantic>=2.6.3
python-docx>=0.8.11

# OCR dependencies
PyMuPDF>=1.19.0
pytesseract>=0.3.8
easyocr>=1.5.0
Pillow>=9.0.0
click>=8.0.0
numpy>=1.20.0
torch>=1.10.0
tqdm>=4.62.0
